create Generation test structure
