+++ Tests/Reqnroll.SystemTests/Generation/GenerationTestBase.cs
@@ -1,0 +1,17 @@
+﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
+
+namespace Reqnroll.SystemTests.Generation;
+
+[TestCategory("Generation")]
+public class GenerationTestBase : SystemTestBase
+{
+    [TestMethod]
+    public void GeneratorAllIn_sample_can_be_handled()
+    {
+        PrepareGeneratorAllInSamples();
+
+        ExecuteTests();
+
+        ShouldAllScenariosPass();
+    }
+}
