+++ Tests/Reqnroll.SystemTests/Generation/MsTestGenerationTest.cs
@@ -1,34 +1,20 @@
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Reqnroll.TestProjectGenerator;
 
-namespace Reqnroll.SystemTests;
+namespace Reqnroll.SystemTests.Generation;
 
 /// <summary>
 /// The purpose of these tests to verify that the tests generated by the MsTest
 /// generator compile and can execute with MsTest.
 /// </summary>
 [TestClass]
-[TestCategory("TestFramework")]
-public class MsTestGeneratorTest : SystemTestBase
+public class MsTestGenerationTest : GenerationTestBase
 {
     protected override void TestInitialize()
     {
         base.TestInitialize();
         _testRunConfiguration.UnitTestProvider = UnitTestProvider.MSTest;
     }
-
-    [TestMethod]
-    public void GeneratorAllIn_sample_can_be_handled()
-    {
-        PrepareGeneratorAllInSamples();
-
-        ExecuteTests();
-
-        ShouldAllScenariosPass();
-    }
-
-    // This is an example of a test that verifies a special case
-
 
     /// <summary>
     /// MsTest v2.* defines the [DataRow] attribute with a ctor that cannot handle if the second
